Update crossref_mapping.fix
This change fixes an issue, that is documented [here](https://github.com/subugoe/goefis/issues/115). This is a typing error (integer instead of string), we 'reuse' `capitalize()` as the Catmandu documentation shows a easy way to do type casts (it seems there is even a concept of data typing missing).