At least provide en error log message when schema validation clashes occur #250
Still most of these errors are programming (or configuration) issues. The end-user
can't do nothing about it except for staring at a screen. The schema validation has the
prower to spot coding mistakes, but can't in its current form hold the business logic
what else apart from the required fields make a record valid.
It is also an exercise in validation. If one makes the validation too strict than users
will get more problems entering metadata. There has to be a balance in what type of errors
are accepted (warnings) and which are really breaking rules (or applications).